requirements.txt
Flask
python-dotenv
firebase-admin
flask-cors
Flask-RESTful
requests 
litellm
openai
gunicorn
google-adk
pinecone==8.1.0